Update AuthController, EmailService, and UserService
