+++ team2/server/.gitignore
@@ -3,6 +3,8 @@
 ##
 ## Get latest from https://github.com/github/gitignore/blob/main/VisualStudio.gitignore
 
+.env
+
 # User-specific files
 *.rsuser
 *.suo
@@ -397,5 +399,4 @@ FodyWeavers.xsd
 # JetBrains Rider
 *.sln.iml
 
-*.json
 *.config
