+++ team2/server/IdeaJarAPI/WebAPI/Services/IEmailService.cs
@@ -1,34 +1,7 @@
-﻿using SendGrid;
-using SendGrid.Helpers.Mail;
-
-namespace WebAPI.Services
+﻿namespace WebAPI.Services
 {
     public interface IEmailService
     {
         Task SendEmail(string toEmail, string subject, string context);
-    }
-
-    public class EmailService : IEmailService
-    {
-        private readonly IConfiguration _configuration;
-        public EmailService(IConfiguration configuration)
-        {
-            _configuration = configuration;
-        }
-
-        public async Task SendEmail(string toEmail,string subject, string content)
-        {
-            //var apiKey = Environment.GetEnvironmentVariable("NAME_OF_THE_ENVIRONMENT_VARIABLE_FOR_YOUR_SENDGRID_KEY");
-            var apiKey = _configuration["SendGrid:ApiKey"];
-            var client = new SendGridClient(apiKey);
-
-            var from = new EmailAddress("ideajar.app@gmail.com", "Idea Jar Support");
-            var to = new EmailAddress(toEmail);
-
-            var msg = MailHelper.CreateSingleEmail(from, to, subject, content, content);
-
-            var response = await client.SendEmailAsync(msg);
-        }
-        
     }
 }
